+++ R/biomassStarter.R
@@ -10,7 +10,7 @@ biomassStarter <- function(x, db, grpBy_quo = NULL, polys = NULL,
   # Read required data and prep the database ------------------------------
   reqTables <- c('PLOT', 'TREE', 'COND', 'POP_PLOT_STRATUM_ASSGN', 
                  'POP_ESTN_UNIT', 'POP_EVAL', 'POP_STRATUM', 'POP_EVAL_TYP', 
-                 'POP_EVAL_GRP')
+                 'POP_EVAL_GRP', 'PLOTGEOM')
 
   # If remote, read in state by state. Otherwise, drop all unnecessary tables.
   db <- readRemoteHelper(x, db, remote, reqTables, nCores)
@@ -75,6 +75,15 @@ biomassStarter <- function(x, db, grpBy_quo = NULL, polys = NULL,
   }
 
   # Other basic variable prep ---------------------------------------------
+  # Join PLOT with PLOTGEOM to allow plot-level geographic attributes to be used
+  # in grpBy statements
+  # First need to get rid of other columns in PLOT in PLOTGEOM. 
+  db$PLOTGEOM <- db$PLOTGEOM %>%
+    dplyr::select(-STATECD, -INVYR, -UNITCD, -COUNTYCD, -PLOT, -LAT, -LON, 
+                  -dplyr::starts_with('CREATED'), -dplyr::starts_with('MODIFIED'))
+  db$PLOT <- db$PLOT %>%
+    dplyr::left_join(db$PLOTGEOM, by = 'CN')
+
   # Get a plot CN and a new pltID that gives a unique ID to each plot
   # PLT_CN is UNITCD, STATECD, COUNTYCD, PLOT, and INVYR
   db$PLOT <- db$PLOT %>%